google_sheets_mcp_lambda/requirements.txt
bedrock-agentcore>=0.1.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
